+++ apps/vitacare-frontend/src/app/globals.css
@@ -29,11 +29,3 @@ body {
   --ring: 212.7 26.8% 83.9%;
 }
 
-body {
-  @apply bg-[#3c3c3c];
-}
-
-@theme{
-  --color-primary:#0096CC;
-}
-
