+++ apps/vitacare-frontend/src/components/atoms/Button/ReceiptButton.tsx
