--- apps/vitacare-frontend/src/components/molecules/README.MD
@@ -1,1 +1,0 @@
-Molecules folder
